measurement Jacobian AIDP for ncamera

+++ include/msckf/InvariantEKF.hpp
@@ -83,7 +83,7 @@ class InvariantEKF : public HybridFilter {
    */
   bool measurementJacobianAIDP(
       const Eigen::Vector4d& ab1rho,
-      const Eigen::Vector2d& obs, int observationIndex,
+      const Eigen::Vector2d& obs, size_t observationIndex,
       std::shared_ptr<const msckf::PointSharedData> pointDataPtr,
       Eigen::Matrix<double, 2, Eigen::Dynamic>* H_x,
       Eigen::Matrix<double, 2, 3>* J_pfi, Eigen::Vector2d* residual) const;
